started adding cancel booking to each row
